refactor: split InsertModals.jsx into focused modules (#79)
Break up the 959-line InsertModals.jsx into per-concern files so each
card type has its own module and the shared chrome lives alongside them.

New files:
- src/components/scene/modals/ModalBase.jsx (shared modal chrome + inputStyle)
- src/components/scene/modals/MemoryCardModal.jsx
- src/components/scene/modals/ImageCardModal.jsx
- src/components/scene/modals/IframeCardModal.jsx
- src/components/scene/modals/TextCardModal.jsx
- src/components/scene/modals/CodeCardModal.jsx
- src/components/scene/modals/VideoCardModal.jsx
- src/components/scene/modals/index.jsx (re-exports)
- src/components/scene/DraggableNumberLabel.jsx
- src/components/scene/ObjectEditPopover.jsx

Updated imports in cardTypes.jsx, DynamicScenePage.jsx, and the
InsertModals tests. Pure mechanical split — no behavior changes.

Closes #79
